Begin work on right (battle) room map 001-2-35
Yellow and red tiles are where monster should not spawn.
Idea is for defeated and observing player to respawn
outside the battle zone.

ToDo: Dynamic disallow re-entry. Going into safe zone is forfeit.
      NPC should warp player back into the arena, conditionally.
      Note documents in evol_akkoteru, or join #evol-dev if this
      is hard to understand.

Included in this commit is a (brutal) "fix" for the S. guard-rail.

Note: This is only the TMX for client. Server may yet need a gat
      or wlk file.