magic-heal
magic-heal-sfx: Originally inma.ogg in my TMW patch.
I assume this is for a casting sound and a lum animation plays.
If the effect lasts for the duration of the heal, then this is
not suitable and a lower vol, shorter sound would be better.

If there is an instant heal type spell, like inma, this should
work and may look fine with lum-like heal, but it's not tested.